requirements.txt
fuzzywuzzy==0.18.0
Levenshtein==0.25.1
lxml==5.2.2
pillow==10.3.0
python-Levenshtein==0.25.1
python-pptx==0.6.23
rapidfuzz==3.9.1
XlsxWriter==3.2.0
